requirements.txt
py-cord~=2.6.1
psutil~=6.1.1
python-dotenv~=1.0.1
uvloop~=0.21.0